requirements.txt
astroid==2.1.0
backcall==0.1.0
certifi==2018.11.29
decorator==4.4.0
Django==2.0.2
django-ckeditor==5.6.1
django-filter==2.2.0
django-js-asset==1.2.2
django-multiselectfield==0.1.8
ipython==7.6.1
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.14.1
lazy-object-proxy==1.3.1
mccabe==0.6.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==5.4.1
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
pylint==2.2.2
pylint-django==2.0.5
pylint-plugin-utils==0.4
pytz==2018.9
six==1.12.0
traitlets==4.3.2
wcwidth==0.1.7
wrapt==1.11.1
